fix: add wrapper to pkg root
