+++ internal/article/delivery/kafka/consumer/consumer.go
@@ -6,7 +6,10 @@ import (
 	articleDomain "github.com/infranyx/go-grpc-template/internal/article/domain"
 	kafkaConsumer "github.com/infranyx/go-grpc-template/pkg/kafka/consumer"
 	"github.com/infranyx/go-grpc-template/pkg/logger"
-	"github.com/infranyx/go-grpc-template/utils/wrapper"
+	"github.com/infranyx/go-grpc-template/pkg/wrapper"
+	wrapperErrorhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/error_handler"
+	wrapperRecoveryhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/recovery_handler"
+	wrapperSentryhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/sentry_handler"
 )
 
 type articleConsumer struct {
@@ -34,9 +37,9 @@ func (ac *articleConsumer) consumerCreateArticle(ctx context.Context, workersNum
 	c := make(chan bool)
 	worker := wrapper.BuildChain(
 		ac.createArticleWorker(ctx, c),
-		wrapper.SentryMiddleware,
-		wrapper.RecoveryMiddleware,
-		wrapper.ErrorHandlerMiddleware,
+		wrapperSentryhandler.SentryHandler,
+		wrapperRecoveryhandler.RecoveryHandler,
+		wrapperErrorhandler.ErrorHandler,
 	)
 	for i := 0; i <= workersNum; i++ {
 		go worker(ctx, nil)
