+++ internal/article/job/job.go
@@ -5,7 +5,10 @@ import (
 
 	articleDomain "github.com/infranyx/go-grpc-template/internal/article/domain"
 	"github.com/infranyx/go-grpc-template/pkg/logger"
-	"github.com/infranyx/go-grpc-template/utils/wrapper"
+	"github.com/infranyx/go-grpc-template/pkg/wrapper"
+	wrapperErrorhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/error_handler"
+	wrapperRecoveryhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/recovery_handler"
+	wrapperSentryhandler "github.com/infranyx/go-grpc-template/pkg/wrapper/handlers/sentry_handler"
 
 	cronJob "github.com/infranyx/go-grpc-template/pkg/cron"
 )
@@ -26,9 +29,9 @@ func (aj *articleJob) RunJobs(ctx context.Context) {
 
 func (aj *articleJob) logArticleJob(ctx context.Context) {
 	worker := wrapper.BuildChain(aj.logArticleWorker(),
-		wrapper.SentryMiddleware,
-		wrapper.RecoveryMiddleware,
-		wrapper.ErrorHandlerMiddleware,
+		wrapperSentryhandler.SentryHandler,
+		wrapperRecoveryhandler.RecoveryHandler,
+		wrapperErrorhandler.ErrorHandler,
 	)
 	entryId, _ := aj.cj.AddFunc("*/1 * * * *",
 		worker.ToCronJobFunc(ctx, nil),
