+++ pkg/wrapper/handlers/recovery_handler/recovery_handler.go
@@ -1,0 +1,25 @@
+package wrapperRecoveryhandler
+
+import (
+	"context"
+
+	"github.com/infranyx/go-grpc-template/pkg/logger"
+	"github.com/infranyx/go-grpc-template/pkg/wrapper"
+	"go.uber.org/zap"
+)
+
+var RecoveryHandler = func(f wrapper.HandlerFunc) wrapper.HandlerFunc {
+	return func(ctx context.Context, args interface{}) (interface{}, error) {
+		defer func() {
+			if r := recover(); r != nil {
+				err, ok := r.(error)
+				if !ok {
+					logger.Zap.Sugar().Errorf("%v", r)
+					return
+				}
+				logger.Zap.Error(err.Error(), zap.Error(err))
+			}
+		}()
+		return f(ctx, args)
+	}
+}
